test(control-box): add a11y test, streamline tests
